Improve handling of stat filtering
This commit is a pretty extensive change to the way filtering is
handled. Previously there was a notion of “direction”, which could have
the values 1 and -1. This approach worked fine for numeric stats but
quickly broke down for non-numeric stats. The concept of direction is
replaced by “filter”, which can have semantic values like
“greaterThanOrEqual” and “notEqual”.  A variety of changes in the
calculations and logic were necessary to support this change.